e9dba924d604ffbc60c2b94ac06ff9e7aefcc0ba Drop model::RescanBlockchain (Abeeujah)
d21add0017ce9a275755a5033b1a2b1c423bb297 Drop model::GetDifficulty (Abeeujah)
1019231a85b05986e149807539b28f827fef4ceb Drop model::GetBlockCount (Abeeujah)

Pull request description:

  Drops the following types in model which do not model a concrete rust-bitcoin type

  - GetBlockCount: newtype on `u64`
  - GetDifficulty: newtype on `f64`
  - RescanBlockchain: both struct fields (start_height, stop_height) are `u32`

ACKs for top commit:
  jamillambert:
    ACK e9dba924d604ffbc60c2b94ac06ff9e7aefcc0ba
  tcharding:
    ACK e9dba924d604ffbc60c2b94ac06ff9e7aefcc0ba

Tree-SHA512: d187f65731a39301190c16f6a40cf9fa883d544ca0debbd452089c3657bbee9d15716e666167b1a508d34ce73c855539ef4e99378cdd1d22487a9309ea6c6260

+++ types/src/v17/wallet/mod.rs
@@ -887,9 +887,9 @@ pub struct LockUnspent(pub bool);
 #[cfg_attr(feature = "serde-deny-unknown-fields", serde(deny_unknown_fields))]
 pub struct RescanBlockchain {
     /// The block height where the rescan has started.
-    pub start_height: i64,
+    pub start_height: u32,
     /// The height of the last rescanned block.
-    pub stop_height: i64,
+    pub stop_height: u32,
 }
 
 /// Result of the JSON-RPC method `sendmany`.
